requirements.txt
# GUI Framework
PyQt6>=6.6.0
PyQt6-QScintilla>=2.14.0
QtAwesome>=1.3.0

# Database Drivers
sqlalchemy>=2.0.0
pymssql>=2.2.0
pyodbc>=5.0.0
mysql-connector-python>=8.0.0
PyMySQL>=1.1.0
psycopg2-binary>=2.9.0
mariadb>=1.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Windows Notifications
windows-toasts>=1.3.0

# Utilities
python-dotenv>=1.0.0
cryptography>=41.0.0
keyring>=24.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
